pyproject.toml
[project]
name = "face-auth-spp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python==4.10.0.84",
    "Flask==3.1.0 ",
    "psycopg2==2.9.10",
    "numpy==2.1.3",
    "python-dotenv==1.0.1",
    "face-recognition==1.3.0",
    "dlib==19.24.6",
    "distribute==0.7.3",
]

